refactor: deprecate some stylable utils
- utils that used deprected stylable/selector utils
- change all stylable utils to use non deprecated helpers
- moved old tests to deprecated folder
- new utils that work with new ast to replace deprecated utils
- copied tests for new utils
- changed scopeNestedSelector to return both ast & selector (immutable)
