+++ packages/core/src/helpers/rule.ts
@@ -1,4 +1,16 @@
+import {
+    parseSelector,
+    stringifySelector,
+    scopeNestedSelector,
+    SelectorNode,
+    walkSelector,
+    convertToSelector,
+    isNodeMatch,
+} from './selector';
+import { getStylableAstData } from './stylable-ast-data';
+import { valueMapping } from '../stylable-value-parsers';
 import * as postcss from 'postcss';
+import cloneDeep from 'lodash.clonedeep';
 
 export function isChildOfAtRule(rule: postcss.Container, atRuleName: string) {
     return (
@@ -45,4 +57,123 @@ export function createWarningRule(
             }),
         ],
     });
+}
+
+export function createSubsetAst<T extends postcss.Root | postcss.AtRule>(
+    root: postcss.Root | postcss.AtRule,
+    selectorPrefix: string,
+    mixinTarget?: T,
+    isRoot = false
+): T {
+    // keyframes on class mixin?
+    const prefixSelectorList = parseSelector(selectorPrefix);
+    const prefixType = prefixSelectorList[0].nodes[0];
+    const containsPrefix = containsMatchInFirstChunk.bind(null, prefixType);
+    const mixinRoot = mixinTarget ? mixinTarget : postcss.root();
+
+    root.nodes.forEach((node) => {
+        if (node.type === `rule`) {
+            const { selectorAst } = getStylableAstData(node);
+            const ast = isRoot
+                ? scopeNestedSelector(prefixSelectorList, selectorAst, true).ast
+                : cloneDeep(selectorAst);
+
+            const matchesSelectors = isRoot ? ast : ast.filter((node) => containsPrefix(node));
+
+            if (matchesSelectors.length) {
+                const selector = stringifySelector(
+                    matchesSelectors.map((selectorNode) => {
+                        // fix chunk ordering
+                        if (!isRoot) {
+                            let startChunkIndex = 0;
+                            let moved = false;
+                            walkSelector(selectorNode, (node, index, nodes) => {
+                                if (node.type === `combinator`) {
+                                    startChunkIndex = index + 1;
+                                    moved = false;
+                                } else if (isNodeMatch(node, prefixType)) {
+                                    if (index > 0 && !moved) {
+                                        moved = true;
+                                        nodes.splice(index, 1);
+                                        nodes.splice(startChunkIndex, 0, node);
+                                    }
+                                }
+                                return undefined;
+                            });
+                        }
+                        // replace prefix
+                        walkSelector(selectorNode, (node) => {
+                            if (isNodeMatch(node, prefixType)) {
+                                convertToSelector(node).nodes = [
+                                    {
+                                        type: `nesting`,
+                                        value: `&`,
+                                        start: node.start,
+                                        end: node.end,
+                                    },
+                                ];
+                            }
+                        });
+                        return selectorNode;
+                    })
+                );
+
+                mixinRoot.append(node.clone({ selector }));
+            }
+        } else if (node.type === `atrule`) {
+            if (node.name === 'media' || node.name === 'supports') {
+                const atRuleSubset = createSubsetAst(
+                    node,
+                    selectorPrefix,
+                    postcss.atRule({
+                        params: node.params,
+                        name: node.name,
+                    }),
+                    isRoot
+                );
+                if (atRuleSubset.nodes) {
+                    mixinRoot.append(atRuleSubset);
+                }
+            } else if (isRoot) {
+                mixinRoot.append(node.clone());
+            }
+        } else {
+            // TODO: add warn?
+        }
+    });
+
+    return mixinRoot as T;
+}
+
+function containsMatchInFirstChunk(prefixType: SelectorNode, selectorNode: SelectorNode) {
+    let isMatch = false;
+    walkSelector(selectorNode, (node) => {
+        if (node.type === `combinator`) {
+            // || node.type === 'spacing') {
+            return walkSelector.stopAll;
+        } else if (node.type === 'pseudo_class') {
+            return walkSelector.skipNested;
+        } else if (isNodeMatch(node, prefixType)) {
+            isMatch = true;
+            return walkSelector.stopAll;
+        }
+        return;
+    });
+    return isMatch;
+}
+
+export function findRule(
+    root: postcss.Root,
+    selector: string,
+    test: any = (statement: any) => statement.prop === valueMapping.extends
+): null | postcss.Declaration {
+    let found: any = null;
+    root.walkRules(selector, (rule) => {
+        const declarationIndex = rule.nodes ? rule.nodes.findIndex(test) : -1;
+        const { isSimpleSelector } = getStylableAstData(rule);
+        if (isSimpleSelector && !!~declarationIndex) {
+            found = rule.nodes[declarationIndex];
+        }
+    });
+    return found;
 }
